Revert "Feature/remove rest client from celery (#728)" (#729)
This reverts commit a1c7c577fcc52902f0c754c6059a36902d0ccab2.
