--- bothub/api/v2/internal/connect_rest_client.py
@@ -1,88 +1,0 @@
-import requests
-from typing import List, Dict
-
-from django.conf import settings
-from django.core.cache import cache
-
-
-class ConnectRESTClient:
-    def __init__(self):
-        self.base_url = settings.CONNECT_API_URL
-        self.headers = {
-            "Content-Type": "application/json; charset: utf-8",
-            "Authorization": self.get_auth_token(),
-        }
-
-    def get_auth_token(self) -> str:
-        request = requests.post(
-            url=settings.OIDC_OP_TOKEN_ENDPOINT,
-            data={
-                "client_id": settings.OIDC_RP_CLIENT_ID,
-                "client_secret": settings.OIDC_RP_CLIENT_SECRET,
-                "grant_type": "client_credentials",
-            },
-        )
-        token = request.json().get("access_token")
-        return f"Bearer {token}"
-
-    def list_classifiers(
-        self, project_uuid: str, user_email: str
-    ) -> List[Dict[str, str]]:
-        cache_key = f"LIST_CLASSIFIER_CACHE_KEY:project_uuid:{project_uuid}:user_email:{user_email}"
-        cache_result = cache.get(cache_key, None)
-        if cache_result is None:
-            request = requests.get(
-                url=f"{self.base_url}/v1/organization/project/list_classifier/",
-                headers=self.headers,
-                params={"project_uuid": project_uuid, "user_email": user_email},
-            )
-            cache.set(cache_key, request.json().get("data"), settings.CACHE_TTL)
-            return request.json().get("data")
-        else:
-            return cache_result
-
-    def list_authorizations(self, project_uuid: str, user_email: str) -> List[str]:
-
-        classifiers = self.list_classifiers(
-            project_uuid=project_uuid, user_email=user_email
-        )
-
-        return [classifier.get("authorization_uuid") for classifier in classifiers]
-
-    def get_authorization_classifier(
-        self, project_uuid: str, authorization_uuid: str, user_email: str
-    ) -> str:
-        """
-        Receives a authorization UUID and returns the respective classifier UUID
-        """
-        classifiers = self.list_classifiers(project_uuid, user_email)
-        classifier = filter(
-            lambda classifier: classifier["authorization_uuid"] == authorization_uuid,
-            classifiers,
-        )
-
-        return next(classifier).get("uuid")
-
-    def remove_authorization(
-        self, project_uuid: str, authorization_uuid: str, user_email: str
-    ):
-        classifier_uuid = self.get_authorization_classifier(
-            project_uuid,
-            authorization_uuid,
-            user_email,
-        )
-        request = requests.delete(
-            url=f"{self.base_url}/v1/organization/project/destroy_classifier/",
-            headers=self.headers,
-            params={"uuid": classifier_uuid, "user_email": user_email},
-        )
-
-        return request.json()
-
-    def create_classifier(self, **kwargs):
-        request = requests.post(
-            url=f"{self.base_url}/v1/organization/project/create_classifier/",
-            headers=self.headers,
-            params={**kwargs, "classifier_type": "bothub"},
-        )
-        return request.json()
