+++ pyproject.toml
@@ -52,7 +52,6 @@ psycopg2-binary = "~=2.9.1"
 weni-protobuffers = "~=1.2.1"
 black = "21.7b0"
 Pillow = "~=8.4.0"
-django-csp = "^3.7"
 
 [tool.poetry.dev-dependencies]
 flake8 = "~=4.0.0"
